+++ ui/cypress/e2e/checks.test.ts
@@ -210,7 +210,9 @@ describe('Checks', () => {
       cy.get('body').tab()
       cy.getByTestID('filter--input checks').should('have.focus')
 
-      cy.focused().tab()
+      cy.focused()
+        .tab()
+        .tab()
       cy.getByTestID('filter--input endpoints').should('have.focus')
 
       cy.focused().tab()