+++ ui/src/shared/apis/queryCache.ts
@@ -20,6 +20,9 @@ import {WINDOW_PERIOD} from 'src/variables/constants'
 export const TIME_INVALIDATION = 1000 * 60 * 10 // 10 minutes
 
 const asSimplyKeyValueVariables = (vari: Variable) => {
+  if (vari.arguments?.type === 'system') {
+    return {[vari.name]: vari.arguments.values || []}
+  }
   return {
     [vari.name]: vari.selected || [],
   }